requirements.txt
openai>=0.27.0
transformers
plotly
beautifulsoup4
qrcode
mysql-connector-python
